basic skeleton designed for language date time
