+++ ner_v1/detectors/temporal/date/constant.py
@@ -1,0 +1,6 @@
+
+
+DATE_CONSTANT_FILE = 'date_constant.csv'
+TIME_CONSTANT_FILE = 'time_constant.csv'
+DATETIME_CONSTANT_FILE = 'datetime_constant.csv'
+NUMERALS_CONSTANT_FILE = 'numbers_constant.csv'
